feat: remove old hooks, keep use-chat-core/operations
