+++ app/components/chat-input/popover-content-auth.tsx
@@ -67,7 +67,7 @@ export function PopoverContentAuth() {
           Login to try more features for free
         </p>
         <p className="text-muted-foreground mb-5 text-base">
-          Add files, use more models, BYO API keys, and more.
+          Add files, use more models, BYOK, and more.
         </p>
         <Button
           variant="secondary"
